==> MovieLens/1-Explore_Datal.ipynb <==
# =================================================
#
# 论题：基于矩阵分解的推荐算法研究
#    1、研究环境：Anaconda3 + Python3.7 + IPython7.12 
#                + JupyterLab + Jupyter Notebook
#    2、数据集：MovieLens(ml-25m)，
#              Kaggle(amazon-ratings)，
#              Book-Crossing Dataset(BX-CSV-Dump)
#    3、研究算法：Funk SVD(LFM)、Bias SVD
#    4、求解方法：随机梯度下降法 SGD
#    5、目标函数：平方损失函数
#    6、评测指标：RMSE、MAE
#
# 思路：数据集探索和模型探索
#    1、初步探索数据集
#    2、评分数据处理，划分数据集和得出稀疏评分矩阵
#    3、LFM 模型探索与模型参数调优
#    4、Bias SVD 模型探索与模型参数调优
#
# 备注：本代码只实用于 MovieLens 数据集，
#       其他两个数据集研究思路和方法一致，
#       只不过是加载不同的数据集而言
# =================================================

# 模块和包都是在逐步的探索中所需要的，全部汇总到这里，
#    并不是一开始就知道了 ^_^ ^_^ ^_^ 
# 不熟悉的模块和包，强烈建议查看官方文档说明以及例子
# 1、导入模块和包
import pandas as pd    # 加载并处理csv文件
import datetime        # 利用datetime处理时间戳
# cPickle 数据以二进制进行高效的储存到文件
import _pickle as cPickle 
# defaultdict 设置稀疏矩阵的 NULL 位置的默认值
from collections import defaultdict 
# 利用scipy sparse 构建稀疏矩阵
import scipy.sparse as ss     
import scipy.io as sio    # 利用scipy储存评分矩阵
# 利用numpy创建指定长度或形状的矩阵以及矩阵运算
import numpy as np 
# numpy.random中的randn函数生成正态分布的随机数据
from numpy.random import random    
import time    # 计算训练时迭代的时间
import json    # 将模型参数保存和加载 json文件
import scipy    # 将储存加载的稀疏评分矩阵转换为numpy矩阵

# 2、加载数据集
# 设置数据储存位置
data_path = "./../dataset/ml-25m/"
# pandas读取文件ratings.csv(两千五百多万行数据)
csv_file = pd.read_csv(data_path + "ratings.csv", sep = ",", nrows=3000000)

# 将时间戳转换
csv_file["timestamp"] = csv_file["timestamp"].map(datetime.datetime.fromtimestamp)
csv_file.head()

# 查看数据维度（行，列）
csv_file.shape

# 将时间戳这个字段属性删除
csv_file = csv_file.drop(['timestamp'], axis=1)
# 查看数据维度（行，列）
csv_file.shape

csv_file.tail()

# 数据集划分训练集和测试集
# =================================================
# Ture 为打乱数据集进行划分，False为不打乱数据集划分
def random_split (df, ratios, shuffle = True):
    
# Function to split DataFrame into train and test
#
# Params:     
#  df (pd.DataFrame): Pandas data frame to be split.
#  ratios (list of floats): list of ratios for split. 
#  The ratios have to sum to 1.
#
# Returns: 
#  list: List of pd.DataFrame split 
#          by the given specifications.
# ###################################################  
    
    seed = 42                  # Set random seed
    if shuffle == True:
        df = df.sample(frac=1)     # Shuffle the data
    samples = df.shape[0]      # Number of samples
    
    # Converts [0.7, 0.2, 0.1] to [0.7, 0.9]
    split_ratio = np.cumsum(ratios).tolist()[:-1] 
    
    # Get the rounded integer split index
    split_index = [round(x * samples) for x in split_ratio]
    
    # split the data
    splits = np.split(df, split_index)
    
    # Add split index 
    # (this makes splitting by group more efficient).
    for i in range(len(ratios)):
        splits[i]["split_index"] = i

    return splits

# 划分数据集
# train,validation,test = random_split()
train, test = random_split(csv_file, [0.8, 0.2])

# 保存数据集为训练集和测试集
# 利用pandas的 DataFrame的to_csv方法
train.to_csv(data_path + "train.csv")
test.to_csv(data_path + "test.csv")

print(test.shape)
print(train.shape)
train.tail()

test.head()
