==> MovieLens/3-LFM_Model.ipynb <==
# 模块和包都是在逐步的探索中所需要的，全部汇总到这里，
#    并不是一开始就知道了 ^_^ ^_^ ^_^ 
# 不熟悉的模块和包，强烈建议查看官方文档说明以及例子
# 1、导入模块和包
import pandas as pd    # 加载并处理csv文件
import datetime        # 利用datetime处理时间戳
# cPickle 数据以二进制进行高效的储存到文件
import _pickle as cPickle 
# defaultdict 设置稀疏矩阵的 NULL 位置的默认值
from collections import defaultdict 
# 利用scipy sparse 构建稀疏矩阵
import scipy.sparse as ss     
import scipy.io as sio    # 利用scipy储存评分矩阵
# 利用numpy创建指定长度或形状的矩阵以及矩阵运算
import numpy as np 
# numpy.random中的randn函数生成正态分布的随机数据
from numpy.random import random    
import time    # 计算训练时迭代的时间
import json    # 将模型参数保存和加载 json文件
import scipy    # 将储存加载的稀疏评分矩阵转换为numpy矩阵

# 4.2.1、LMF模型
# ############################################
#
# 核心算法实现
#
# @输入参数
#     R —— M*N 评分矩阵
#     k —— 隐向量的维度
#     theta —— 迭代次数
#     alpha —— 步长（学习率）
#     lamda —— 正则化系数
#
# @输出参数
#     分解之后的 P，Q
#     P：初始化用户特征矩阵 M*K
#     Q：初始化物品特征矩阵 N*K
#
# ############################################
# 设定模型参数
# 18 个电影类型 10、12、14、16、18、20、24、36
K = 16
# 迭代次数：20、40、60、80、100
theta = 80
alpha = 0.005
lamda = 0.15
# 核心算法
def LFM_grad_desc( R, K, theta, alpha, lamda ):
    # 基本维度参数定义
    M = R.shape[0]
    N = R.shape[1]
    
    # P,Q初始值，随机生成
    P = np.random.rand(M, K)
    Q = np.random.rand(N, K)
    Q = Q.T
    
    # 第一次迭代时间开始
    time_start = time.time()
    print("开始进行{}个step的训练".format(theta))
    each_time_start = time_start
        
    # 开始迭代
    for step in range(theta):
        print('The {}-th  step is running'.format(step))
        # 对所有的用户u、物品i做遍历，
        # 对应的特征向量Pu、Qi梯度下降
        for u in range(M):
            for i in range(N):
                # 对于每一个大于0的评分，
                # 求出预测评分误差
                if R[u, i] > 0:
                    eui = np.dot( P[u,:], Q[:,i] ) - R[u, i]
                    
                    # 按照梯度下降算法更新当前的Pu、Qi
                    for k in range(K):
                        P[u][k] = P[u][k] - alpha * ( eui * Q[k][i] + lamda * P[u][k] )
                        Q[k][i] = Q[k][i] - alpha * ( eui * P[u][k] + lamda * Q[k][i] )
        
        # u、i遍历完成，所有特征向量更新完成，
        # 可以得到P、Q，可以计算预测评分矩阵
        predR = np.dot( P, Q )
        
        # 计算当前损失函数
        cost = 0
        for u in range(M):
            for i in range(N):
                if R[u, i] > 0:
                    cost += ( np.dot( P[u,:], Q[:,i] ) - R[u, i] ) ** 2
                    # 加上正则化项
                    for k in range(K):
                        cost += lamda * ( P[u][k] ** 2 + Q[k][i] ** 2 )
                        each_cost = np.sqrt(cost / K)
        
        # 学习率递减 / learning rate improve 3th
        alpha = alpha * 0.93
        # 每次迭代时间结束
        each_time_tick = time.time()
        # 每次迭代消耗的时间
        each_cost_time = each_time_tick - each_time_start
        # 更新计算每次迭代的时间
        each_time_start = each_time_tick
        print("完成第{}个step的训练, each_cost={}, 耗时{:.4f}秒".format( step + 1, cost, each_cost_time))
        
        if cost < 0.0001:
            break
    # 计算训练数据集消耗的总时间
    time_end = time.time()
    total_cost_time = time_end - time_start
    print("结束了{}个step的训练，总耗时{:.4f}秒".format(theta, total_cost_time))    
    return P, Q.T, cost

# 加载评分矩阵
data_path = "./../dataset/ml-25m/"
user_item_score = sio.mmread(data_path + "user_item_score")
# todense() 转换为矩阵 numpy 
R = scipy.sparse.csc_matrix.todense(user_item_score)
type(R)

# 开始训练 LFM 模型
P, Q, ess = LFM_grad_desc( R, K, theta, alpha, lamda )

# 通过训练的 P和 Q 计算出预测评分矩阵
pred_R = np.dot( P, Q.T )
# 将预测评分矩阵保存
sio.mmwrite(data_path + "pred_R", pred_R)

# 查看原始评分矩阵
print(R)
# 查看预测评分矩阵
print(pred_R)

# 查看训练后的 P 和 Q 矩阵
print(P)
print("====================================================")
print(Q)

# LFM模型评测指标
# RMSE 和 MAE
def calc_evaluation(R, pred_R):
    mae_sum = 0.0
    rmse_sum = 0.0
    # 基本维度参数定义
    M = R.shape[0]
    N = R.shape[1]
    T = 0
    
    for user in range(M):
        for item in range(N):
            if R[user, item] > 0:
                T += 1
                ess = R[user, item] - pred_R[user, item]
                mae_sum += abs(ess)
                rmse_sum += ess**2
    
    # 计算总的 MAE 指标
    MAE = mae_sum / T
    # Root Mean Square Error RMSE
    RMSE = np.sqrt(rmse_sum / T)
    return MAE, RMSE

MAE, RMSE = calc_evaluation(R, pred_R)
print("the MAE of LFM:", MAE)
print("the RMSE OF LFM:", RMSE)
